refactor: use new index exports across frontend
