+++ frontend/src/components/admin/adminLayout.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Outlet, Navigate } from 'react-router-dom';
-import { useAuth } from '../../hooks/useAuth';
+import { useAuth } from '../../hooks';
 import AdminSidebar from './AdminSidebar';
 import AdminHeader from './AdminHeader';
