+++ frontend/src/components/auth/forgotPasswordForm.jsx
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
 import { Link } from 'react-router-dom';
-import { useForm } from '../../hooks/useForm';
+import { useForm } from '../../hooks';
 import AuthAPI from '../../api/auth';
 import Input from '../../common/Input';
 import Button from '../../common/Button';
