+++ frontend/src/components/common/navbar.jsx
@@ -1,7 +1,6 @@
 import React, { useState, useEffect } from 'react';
 import { Link, useLocation, useNavigate } from 'react-router-dom';
-import { useAuth } from '../../hooks/useAuth';
-import { useNotifications } from '../../hooks/useNotifications';
+import { useAuth, useNotifications } from '../../hooks';
 import Button from './Button';
 
 const Navbar = () => {
