+++ frontend/src/components/referrals/index.js
@@ -1,0 +1,3 @@
+export { default as ReferralInstructions } from './ReferralInstructions';
+export { default as ReferralList } from './ReferralList';
+export { default as ReferralHistory } from './ReferralHistory';
