+++ frontend/src/layouts/dashboardLayout.jsx
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
 import { Outlet, Link, useLocation, Navigate } from 'react-router-dom';
-import { useAuth } from '../hooks/useAuth';
+import { useAuth } from '../hooks';
 import Navbar from '../components/common/Navbar';
 import Footer from '../components/common/Footer';
