+++ frontend/src/pages/auth/forgotPassword.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import ForgotPasswordForm from '../../components/auth/ForgotPasswordForm';
+import { ForgotPasswordForm } from '../../components/auth';
 
 const ForgotPassword = () => {
   return <ForgotPasswordForm />;
