+++ frontend/src/pages/auth/login.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import LoginForm from '../../components/auth/LoginForm';
+import { LoginForm } from '../../components/auth';
 
 const Login = () => {
   return <LoginForm />;
