+++ frontend/src/pages/auth/register.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import RegisterForm from '../../components/auth/RegisterForm';
+import { RegisterForm } from '../../components/auth';
 
 const Register = () => {
   return <RegisterForm />;
