+++ frontend/src/pages/referralsPage.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
-import { useAuth } from '../hooks/useAuth';
+import { useAuth } from '../hooks';
 import { Navigate } from 'react-router-dom';
-import ReferralList from '../components/referrals/ReferralList';
+import { ReferralList } from '../components/referrals';
 
 const ReferralsPage = () => {
   const { user } = useAuth();
