+++ frontend/src/utils/index.js
@@ -1,0 +1,7 @@
+const anonymizer = require('./anonymizer');
+const errorHandler = require('./error-handler');
+
+module.exports = {
+  ...anonymizer,
+  errorHandler
+};
